third_party/blink/web_tests/virtual/webnn-service-on-cpu/external/wpt/webnn/idlharness.https.any-expected.txt

This is a testharness.js-based test.
Found 4 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] MLContext interface: operation compute(MLGraph, MLNamedArrayBufferViews, MLNamedArrayBufferViews)
  assert_own_property: interface prototype object missing non-static operation expected property "compute" missing
[FAIL] MLContext interface: context must inherit property "compute(MLGraph, MLNamedArrayBufferViews, MLNamedArrayBufferViews)" with the proper type
  assert_inherits: property "compute" not found in prototype chain
[FAIL] MLContext interface: calling compute(MLGraph, MLNamedArrayBufferViews, MLNamedArrayBufferViews) on context with too few arguments must throw TypeError
  assert_inherits: property "compute" not found in prototype chain
[FAIL] MLGraphBuilder interface: operation softmax(MLOperand, unsigned long, optional MLOperatorOptions)
  assert_equals: property has wrong .length expected 2 but got 1
Harness: the test ran to completion.
